requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==2.5.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
chromedriver-autoinstaller==0.3.1
cryptography==37.0.2
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
h11==0.13.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==4.6.1
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
outcome==1.2.0
parsimonious==0.8.1
protobuf==3.20.1
pybrowsers==0.4.1
pycparser==2.21
pycryptodome==3.15.0
pyOpenSSL==22.0.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dotenv==0.20.0
pywin32==303
requests==2.28.1
rlp==2.0.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toolz==0.11.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
varint==1.0.2
web3==5.29.2
webdriver-manager==3.8.0
websockets==9.1
wsproto==1.1.0
yarl==1.7.2
